add all binaries
