+++ .gitignore
@@ -5,6 +5,7 @@ integration_tests/recorder-extension/node_modules/
 integration_tests/recorder-extension/ext.zip
 integration_tests/.DS_Store
 integration_tests/bootstrap
+integration_tests/src/bin
 .DS_Store
 .layers
 extensions
