+++ integration_tests/snapshots/enhanced-metric-go
@@ -1,0 +1,9 @@
+[sketch] {"metric":"aws.lambda.enhanced.billed_duration","distributions":null,"tags":["account_id:601427279990","architecture:x86_64","aws_account:601427279990","cold_start:true","dd_extension_version:123","env:integration-tests-env","function_arn:arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-enhanced-metric-go","functionname:integration-tests-extension-XXXXXX-enhanced-metric-go","memorysize:1024","region:sa-east-1","resource:integration-tests-extension-XXXXXX-enhanced-metric-go","service:integration-tests-service","taga:valuea","tagb:valueb","tagc:valuec","tagd:valued","version:integration-tests-version"],"dogsketches":[{"ts":XXX,"cnt":1,"min":XXX,"max":XXX,"avg":XXX,"sum":XXX,"k":[XXX],"n":[1]}]} 
+[sketch] {"metric":"aws.lambda.enhanced.duration","distributions":null,"tags":["account_id:601427279990","architecture:x86_64","aws_account:601427279990","cold_start:true","dd_extension_version:123","env:integration-tests-env","function_arn:arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-enhanced-metric-go","functionname:integration-tests-extension-XXXXXX-enhanced-metric-go","memorysize:1024","region:sa-east-1","resource:integration-tests-extension-XXXXXX-enhanced-metric-go","service:integration-tests-service","taga:valuea","tagb:valueb","tagc:valuec","tagd:valued","version:integration-tests-version"],"dogsketches":[{"ts":XXX,"cnt":1,"min":XXX,"max":XXX,"avg":XXX,"sum":XXX,"k":[XXX],"n":[1]}]} 
+[sketch] {"metric":"aws.lambda.enhanced.estimated_cost","distributions":null,"tags":["account_id:601427279990","architecture:x86_64","aws_account:601427279990","cold_start:true","dd_extension_version:123","env:integration-tests-env","function_arn:arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-enhanced-metric-go","functionname:integration-tests-extension-XXXXXX-enhanced-metric-go","memorysize:1024","region:sa-east-1","resource:integration-tests-extension-XXXXXX-enhanced-metric-go","service:integration-tests-service","taga:valuea","tagb:valueb","tagc:valuec","tagd:valued","version:integration-tests-version"],"dogsketches":[{"ts":XXX,"cnt":1,"min":XXX,"max":XXX,"avg":XXX,"sum":XXX,"k":[XXX],"n":[1]}]} 
+[sketch] {"metric":"aws.lambda.enhanced.init_duration","distributions":null,"tags":["account_id:601427279990","architecture:x86_64","aws_account:601427279990","cold_start:true","dd_extension_version:123","env:integration-tests-env","function_arn:arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-enhanced-metric-go","functionname:integration-tests-extension-XXXXXX-enhanced-metric-go","memorysize:1024","region:sa-east-1","resource:integration-tests-extension-XXXXXX-enhanced-metric-go","service:integration-tests-service","taga:valuea","tagb:valueb","tagc:valuec","tagd:valued","version:integration-tests-version"],"dogsketches":[{"ts":XXX,"cnt":1,"min":XXX,"max":XXX,"avg":XXX,"sum":XXX,"k":[XXX],"n":[1]}]} 
+[sketch] {"metric":"aws.lambda.enhanced.invocations","distributions":null,"tags":["account_id:601427279990","architecture:x86_64","aws_account:601427279990","cold_start:true","datadog_lambda:vX.X.X","dd_extension_version:123","dd_lambda_layer:datadog-go1.17","env:integration-tests-env","function_arn:arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-enhanced-metric-go","functionname:integration-tests-extension-XXXXXX-enhanced-metric-go","memorysize:1024","region:sa-east-1","resource:integration-tests-extension-XXXXXX-enhanced-metric-go","service:integration-tests-service","taga:valuea","tagb:valueb","tagc:valuec","tagd:valued","version:integration-tests-version"],"dogsketches":[{"ts":XXX,"cnt":1,"min":1,"max":1,"avg":1,"sum":1,"k":[XXX],"n":[1]}]} 
+[sketch] {"metric":"aws.lambda.enhanced.max_memory_used","distributions":null,"tags":["account_id:601427279990","architecture:x86_64","aws_account:601427279990","cold_start:true","dd_extension_version:123","env:integration-tests-env","function_arn:arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-enhanced-metric-go","functionname:integration-tests-extension-XXXXXX-enhanced-metric-go","memorysize:1024","region:sa-east-1","resource:integration-tests-extension-XXXXXX-enhanced-metric-go","service:integration-tests-service","taga:valuea","tagb:valueb","tagc:valuec","tagd:valued","version:integration-tests-version"],"dogsketches":[{"ts":XXX,"cnt":1,"min":XXX,"max":XXX,"avg":XXX,"sum":XXX,"k":[XXX],"n":[1]}]} 
+[sketch] {"metric":"aws.lambda.enhanced.memorysize","distributions":null,"tags":["account_id:601427279990","architecture:x86_64","aws_account:601427279990","cold_start:true","dd_extension_version:123","env:integration-tests-env","function_arn:arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-enhanced-metric-go","functionname:integration-tests-extension-XXXXXX-enhanced-metric-go","memorysize:1024","region:sa-east-1","resource:integration-tests-extension-XXXXXX-enhanced-metric-go","service:integration-tests-service","taga:valuea","tagb:valueb","tagc:valuec","tagd:valued","version:integration-tests-version"],"dogsketches":[{"ts":XXX,"cnt":1,"min":XXX,"max":XXX,"avg":XXX,"sum":XXX,"k":[XXX],"n":[1]}]} 
+[sketch] {"metric":"aws.lambda.enhanced.runtime_duration","distributions":null,"tags":["account_id:601427279990","architecture:x86_64","aws_account:601427279990","cold_start:true","dd_extension_version:123","env:integration-tests-env","function_arn:arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-enhanced-metric-go","functionname:integration-tests-extension-XXXXXX-enhanced-metric-go","memorysize:1024","region:sa-east-1","resource:integration-tests-extension-XXXXXX-enhanced-metric-go","service:integration-tests-service","taga:valuea","tagb:valueb","tagc:valuec","tagd:valued","version:integration-tests-version"],"dogsketches":[{"ts":XXX,"cnt":1,"min":XXX,"max":XXX,"avg":XXX,"sum":XXX,"k":[XXX],"n":[1]}]} 
+[sketch] {"metric":"serverless.lambda-extension.integration-test.count","distributions":null,"tags":["account_id:601427279990","architecture:x86_64","aws_account:601427279990","dd_extension_version:123","dd_lambda_layer:datadog-go1.17","env:integration-tests-env","function_arn:arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-enhanced-metric-go","functionname:integration-tests-extension-XXXXXX-enhanced-metric-go","memorysize:1024","region:sa-east-1","resource:integration-tests-extension-XXXXXX-enhanced-metric-go","service:integration-tests-service","taga:valuea","tagb:valueb","tagc:valuec","tagd:valued","version:integration-tests-version"],"dogsketches":[{"ts":XXX,"cnt":1,"min":1,"max":1,"avg":1,"sum":1,"k":[XXX],"n":[1]}]}
