+++ integration_tests/snapshots/no-enhanced-metric-go
@@ -1,0 +1,1 @@
+
