+++ integration_tests/snapshots/simple-trace-python
@@ -1,1 +1,2 @@
+[trace] {"traceID":XXX,"spans":[{"service":"aws.lambda","name":"aws.lambda","resource":"integration-tests-extension-XXXXXX-simple-trace-python","trace_id":XXX,"span_id":XXX,"parent_id":XXX,"start":XXX,"duration":XXX,"error":0,"meta":{"_dd.compute_stats":"1","_dd.origin":"lambda","account_id":"601427279990","architecture":"x86_64","aws_account":"601427279990","cold_start":"false","datadog_lambda":"X.X.X","dd_extension_version":"123","dd_trace":"X.X.X","env":"integration-tests-env","function_arn":"arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-simple-trace-python","function_version":"$LATEST","functionname":"integration-tests-extension-XXXXXX-simple-trace-python","memorysize":"1024","region":"sa-east-1","request_id":"XXX","resource_names":"integration-tests-extension-XXXXXX-simple-trace-python","runtime-id":"XXX","service":"integration-tests-service","tagA":"valueA","tagB":"valueB","taga":"valuea","tagb":"valueb","tagc":"valuec","tagd":"valued","version":"integration-tests-version"},"metrics":{"_dd.agent_psr":1,"_dd.tracer_kr":1,"_sampling_priority_v1":1,"_top_level":1,"system.pid":XXX},"type":"serverless"},{"service":"aws.lambda","name":"integration_test","resource":"integration-test","trace_id":XXX,"span_id":XXX,"parent_id":XXX,"start":XXX,"duration":XXX,"error":0,"meta":{"_dd.compute_stats":"1","_dd.origin":"lambda","account_id":"601427279990","architecture":"x86_64","aws_account":"601427279990","dd_extension_version":"123","env":"integration-tests-env","function_arn":"arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-simple-trace-python","functionname":"integration-tests-extension-XXXXXX-simple-trace-python","memorysize":"1024","region":"sa-east-1","service":"integration-tests-service","tagA":"valueA","tagB":"valueB","taga":"valuea","tagb":"valueb","tagc":"valuec","tagd":"valued","version":"integration-tests-version"},"metrics":null,"type":""}],"startTime":XXX,"endTime":XXX}
 [trace] {"traceID":XXX,"spans":[{"service":"aws.lambda","name":"aws.lambda","resource":"integration-tests-extension-XXXXXX-simple-trace-python","trace_id":XXX,"span_id":XXX,"parent_id":XXX,"start":XXX,"duration":XXX,"error":0,"meta":{"_dd.compute_stats":"1","_dd.origin":"lambda","account_id":"601427279990","architecture":"x86_64","aws_account":"601427279990","cold_start":"true","datadog_lambda":"X.X.X","dd_extension_version":"123","dd_trace":"X.X.X","env":"integration-tests-env","function_arn":"arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-simple-trace-python","function_version":"$LATEST","functionname":"integration-tests-extension-XXXXXX-simple-trace-python","memorysize":"1024","region":"sa-east-1","request_id":"XXX","resource_names":"integration-tests-extension-XXXXXX-simple-trace-python","runtime-id":"XXX","service":"integration-tests-service","tagA":"valueA","tagB":"valueB","taga":"valuea","tagb":"valueb","tagc":"valuec","tagd":"valued","version":"integration-tests-version"},"metrics":{"_dd.agent_psr":1,"_dd.tracer_kr":1,"_sampling_priority_v1":1,"_top_level":1,"system.pid":XXX},"type":"serverless"},{"service":"aws.lambda","name":"integration_test","resource":"integration-test","trace_id":XXX,"span_id":XXX,"parent_id":XXX,"start":XXX,"duration":XXX,"error":0,"meta":{"_dd.compute_stats":"1","_dd.origin":"lambda","account_id":"601427279990","architecture":"x86_64","aws_account":"601427279990","dd_extension_version":"123","env":"integration-tests-env","function_arn":"arn:aws:lambda:sa-east-1:601427279990:function:integration-tests-extension-XXXXXX-simple-trace-python","functionname":"integration-tests-extension-XXXXXX-simple-trace-python","memorysize":"1024","region":"sa-east-1","service":"integration-tests-service","tagA":"valueA","tagB":"valueB","taga":"valuea","tagb":"valueb","tagc":"valuec","tagd":"valued","version":"integration-tests-version"},"metrics":null,"type":""}],"startTime":XXX,"endTime":XXX}
