+++ integration_tests/src/go-tests/without-ddlambda/main.go
@@ -1,0 +1,24 @@
+package main
+
+import (
+	"context"
+
+	"github.com/aws/aws-lambda-go/events"
+	"github.com/aws/aws-lambda-go/lambda"
+)
+
+type testResponse struct {
+	StatusCode int    `json:"statusCode"`
+	Body       string `json:"body"`
+}
+
+func PingHandler(ctx context.Context, ev events.APIGatewayProxyRequest) (testResponse, error) {
+	return testResponse{
+		StatusCode: 200,
+		Body:       "ok",
+	}, nil
+}
+
+func main() {
+	lambda.Start(PingHandler)
+}
