--- integration_tests/src/traceGo.go
@@ -1,41 +1,0 @@
-package main
-
-import (
-	"context"
-	"net/http"
-	"time"
-
-	"github.com/aws/aws-lambda-go/lambda"
-
-	ddlambda "github.com/DataDog/datadog-lambda-go"
-	"github.com/aws/aws-lambda-go/events"
-	httptrace "gopkg.in/DataDog/dd-trace-go.v1/contrib/net/http"
-	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
-)
-
-type testResponse struct {
-	StatusCode int    `json:"statusCode"`
-	Body       string `json:"body"`
-}
-
-func handleRequest(ctx context.Context, ev events.APIGatewayProxyRequest) (testResponse, error) {
-
-	req, _ := http.NewRequestWithContext(ctx, "GET", "https://www.datadoghq.com", nil)
-	client := http.Client{}
-	client = *httptrace.WrapClient(&client)
-	client.Do(req)
-
-	// Create a custom span
-	s, _ := tracer.StartSpanFromContext(ctx, "child.span")
-	time.Sleep(100 * time.Millisecond)
-	s.Finish()
-
-	return testResponse{
-		StatusCode: 200,
-		Body:       "ok",
-	}, nil
-}
-
-func main() {
-	lambda.Start(ddlambda.WrapHandler(handleRequest, nil))
-}
